Fully working tests
