+++ requirements_test.txt
@@ -2,3 +2,4 @@ docker >= 5.0.3
 netifaces >= 0.11.0
 mysql-connector-python==8.0.29
 pytest
+pytest-subtests
